requirements.txt
streamlit
pymysql
scikit-learn
numpy
intel-openmp
idna
pillow
pandas